update :: percent 합계 기댓값 100으로 변경

+++ src/main/java/team/startup/expo/domain/stat/presentation/dto/response/StandardParticipantStatsResponse.java
@@ -16,34 +16,34 @@ public class StandardParticipantStatsResponse {
     @Builder
     public static class ElementaryDto {
         private Integer number;
-        private Integer percent;
+        private Float percent;
     }
 
     @Getter
     @Builder
     public static class MiddleDto {
         private Integer number;
-        private Integer percent;
+        private Float percent;
     }
 
     @Getter
     @Builder
     public static class HighDto {
         private Integer number;
-        private Integer percent;
+        private Float percent;
     }
 
     @Getter
     @Builder
     public static class KindergartenDto {
         private Integer number;
-        private Integer percent;
+        private Float percent;
     }
 
     @Getter
     @Builder
     public static class OtherDto {
         private Integer number;
-        private Integer percent;
+        private Float percent;
     }
 }
